refactor: move search functions to utils.py
